refactor: post, comment 공통 로직 및 필드 CommunityBaseEntity로 추상화
